优化热灸页面 实现控制温度 时间 等逻辑判断

+++ NBlue/ViewController/View/HealthMangerView/CalenderUIView/NLCalenderPicker.m
@@ -127,7 +127,7 @@ static const NSInteger BTNTAG = 4000;
 }
 -(void)pickerView:(UIPickerView *)pickerView didSelectRow:(NSInteger)row inComponent:(NSInteger)component{
     _choicePickerIndex = row;
-    NSLog(@"%ld",row);
+    NSLog(@"%ld",(long)row);
 }
 #pragma mark 自己的代理
 #pragma mark 按钮事件
